CRUD(edit category)

+++ DiplomToyStore/Domain/AbstractRepo/IProductRepository.cs
@@ -6,7 +6,7 @@ namespace DiplomToyStore.Domain.AbstractRepo
     public interface IProductRepository
     {
         IEnumerable<Product> Products { get; }
-        Product GetProductById(long id);
+        Product GetProductById(int id);
         void AddProduct(Product product);
         void UpdateProduct(Product product);
         void DeleteProduct(int id);
